+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: Fix `burnAsset` return value and types
+
 ## 0.2.2 (2025-09-09)
 
 - added: Add `burnAsset` method